+++ families/mc1201/common/src/main/java/dev/foucaultleon/flterraforged/minecraft/mc1201/worldgen/FlTerraForgedChunkGenerator.java
@@ -267,7 +267,7 @@ public final class FlTerraForgedChunkGenerator extends ChunkGenerator {
 
     private TerrainWorld bind(NoiseConfig noiseConfig) {
         TerrainWorld world = session.bind(noiseConfig);
-        engineBiomeSource.bind(world);
+        engineBiomeSource.bind(world, getSeaLevel());
         return world;
     }